Only remove the existing `serai-runtime` build directory when not doing an incremental build
This advertently forced every build to be from scratch, disabling all benefits
of incremental builds, when the intent was solely not to keep old build caches
when incremental builds were disabled and wouldn't be of benefit.

Removes a couple minutes from `clippy` (which does build `serai-runtime`'s
WASM). In the future, we should investigate if `clippy` (which can unofficially
be detected via `RUSTC_WORKSPACE_WRAPPER`) should short-circuit any behavior
present.

+++ substrate/runtime/build.rs
@@ -567,8 +567,13 @@ which will build the WASM as part of its build process, with the necessary confi
     will be locked by the build process we're currently running.
   */
   let target_dir = PathBuf::from(cargo_env("OUT_DIR")).join("target");
-  // Remove the directory if it already exists
-  let _ = fs::remove_dir_all(&target_dir);
+  if release_wasm() {
+    /*
+      Remove the directory if it already exists, as this will be a non-incremental build so at best
+      it does nothing, and at worst it accumulates due to `cargo`'s lack-luster garbage collection.
+    */
+    let _ = fs::remove_dir_all(&target_dir);
+  }
   build_command.env("CARGO_TARGET_DIR", &target_dir);
 
   /*
